fix warnings

+++ tests/test_json_serializer.cpp
@@ -172,7 +172,7 @@ TEST(jsonSerializer, binary2)
 
     const BinData2 expected = { std::make_unique<int16_t[]>(expected.size), buf1 };
 
-    for (int16_t i = 0; i < expected.size; ++i)
+    for (size_t i = 0; i < expected.size; ++i)
     {
         expected.x.get()[i] = i + 1;
         expected.y[i] = i + 10;
@@ -190,7 +190,7 @@ TEST(jsonSerializer, binary2)
     pods::JsonDeserializer<pods::ReadOnlyMemoryStorage> deserializer(in);
     EXPECT_EQ(deserializer.load(actual), pods::Error::NoError);
 
-    for (int16_t i = 0; i < expected.size; ++i)
+    for (size_t i = 0; i < expected.size; ++i)
     {
         EXPECT_EQ(actual.x.get()[i], i + 1);
         EXPECT_EQ(actual.y[i], i + 10);
